+++ cedar-testing/tests/cedar-policy-cli/decimal.rs
@@ -14,31 +14,22 @@
  * limitations under the License.
  */
 
-//! Integration tests targeting the ipaddr extension
-//!
+//! Integration tests targeting the decimal extension
 
-use cedar_policy::integration_testing::perform_integration_test_from_json;
+use super::perform_integration_test_from_json;
 use std::path::Path;
 
 /// Path of the folder containing the JSON tests
 fn folder() -> &'static Path {
-    Path::new("tests/ip")
+    Path::new("tests/decimal")
 }
 
 #[test]
-#[cfg(feature = "ipaddr")]
-fn ip_1() {
+fn decimal_1() {
     perform_integration_test_from_json(folder().join("1.json"));
 }
 
 #[test]
-#[cfg(feature = "ipaddr")]
-fn ip_2() {
+fn decimal_2() {
     perform_integration_test_from_json(folder().join("2.json"));
-}
-
-#[test]
-#[cfg(feature = "ipaddr")]
-fn ip_3() {
-    perform_integration_test_from_json(folder().join("3.json"));
 }